requirements.txt
streamlit>=1.45.1
google-generativeai>=0.8.5
langchain>=0.3.25
langchain-community>=0.3.24
langchain-google-genai>=2.0.10
chromadb>=1.0.11
pypdf>=5.5.0
sentence-transformers>=4.1.0
python-dotenv>=1.1.0
